Problem at the start of a new day
There was a problem where the client got stuck after the sales menu.
There is no problem with selling. However, the error could be triggered
when someone sold something. The command `Game1.getFarm().mailbox();`
was called with each new day. Normally, it opens the messages in the
mailbox. In the handler `helper.Events.GameLoop.DayStarted`, the dialog
is simply closed. However, there is no confirmation for the client, who
then waits forever. The command was transferred to the behavior chain as
the `CheckTheMailboxBehaviorLink` class. The opened messages are handled
by the `ProcessDialogueBehaviorLink` class in the `LetterViewerMenu`
block.